crypto: hctr2 - Add HCTR2 support
Add support for HCTR2 as a template. HCTR2 is a wide-block cipher mode, meaning
that it has stronger security properties than other encryption modes.

For example, with XTS mode, a single bitflip in the plaintext only
changes 16 bytes in the ciphertext. With HCTR2 mode, a single bitflip
changes the entire ciphertext to look completely random. Moreover, HCTR2
allows IVs to be reused without the risk of correlated ciphertexts. This
makes HCTR2 especially useful for disk encryption and filename
encryption.

HCTR2 requires POLYVAL and XCTR as components. More information on HCTR2 can be
found here: Length-preserving encryption with HCTR2:
https://eprint.iacr.org/2021/1441.pdf

Signed-off-by: Nathan Huckleberry <nhuck@google.com>