Support CONFIG_BLK_IN_WAITER
This patch adds an API for block device drivers and filesystems to
perform post-processing on bios. This API submits the post-processing
work to any IO waiters and the specified workqueue at the same time.
Whichever handler wakes first will handle the work.

This can significantly improve latency for drivers which require
intensive IO post-processing for the following reasons:
1. When processing in-waiter, the scheduler is only invoked once. If the
   IO device is fast enough, the scheduler may not be invoked at all.
   Directly using a workqueue requires at least two scheduler calls.
2. Asynchronous IO can no longer block synchronous IO in the workqueue.
3. The reading task's priority is bound to the post-processing job.
4. Jobs can go around the workqueue allowing for greater parallelism.

If the waiting code does not support in-waiter post-processing or if
there are no waiters, the post-processing defaults to the specified
workqueue.

Change-Id: I78fa9181402dde6b6828e5e1753012f96c331fd4