Allow IO post-processing in waiter context.
This patchset introduces a function called bio_defer_work() which is
designed to improve the post-processing of disk IO.  It aims to reduce
latency and address several issues with the current approaches of using
a workqueue or performing post-processing in interrupt context.

bio_defer_work() allows post-processing to be done in the task context
of any IO waiters, which reduces the number of scheduler calls,
separates asynchronous and synchronous workloads, and allows for greater
concurrency. The function has been tested with the dm-verity driver on a
ramdisk and a higher latency disk, and has shown a reduction in latency
of around 30 microseconds.

This function can be used to replace calls to queue_work() and may be
useful for improving the latency of post-processing in drivers such as
dm-crypt, dm-verity, fscrypt, fsverity, f2fs decompression, and erofs
decompression.

The current primary approach for intensive post-processing involves
deferring the IO post-processing to a workqueue. This has several
disadvantages:
  1. The scheduler must run twice. Once to wake the workqueue and once
     to wake the waiting process. This adds unnecessary latency. When
     stacking drivers which require post-processing, this latency stacks
     as well.
  2. Asynchronous IO is queued in the same workqueue as synchronous IO.
     These asynchronous requests must be handled before synchronous
     requests.
  3. All post-processing of the same type is handled with the same
     priority.  IO from high priority processes may be blocked by IO
     from low priority processes.
  4. Multiple threads performing IO at once can be bottlenecked by the
     workqueue.
  5. The verification step does not contribute to the original task’s
     timeslice. This may make scheduling decisions suboptimal.

A secondary approach, which several drivers have adopted, involves
performing post-processing directly in interrupt context. This approach
gives a performance increase, but causes jitter for real-time threads
and can only be used for post-processing that does not sleep.

This proposed third approach aims to improve performance by skipping a
scheduler call and prioritizing post-processing based on the reading
task's priority. This approach also allows synchronous IO to take
precedence over asynchronous IO, avoids impact on real-time threads and
does not have the limitations of the in-interrupt approach.

This code has been tested using dm-verity on a ramdisk and on an ssd.

dm-verity on a ramdisk:
Before:
    clat (usec): min=10, max=11810, avg=45.70, stdev=92.28
    READ: bw=81.2MiB/s, io=512MiB, run=6309-6309msec
After:
    clat (usec): min=5, max=7915, avg=16.05, stdev=61.18
    READ: bw=220MiB/s, io=512MiB, run=2330-2330msec
dm-verity on a higher latency disk:
Before:
    clat (usec): min=20, max=7818, avg=120.29, stdev=42.58
    READ: bw=31.6MiB/s, io=949MiB, run=30001-30001msec
After:
    clat (usec): min=20, max=9508, avg=96.08, stdev=30.00
    READ: bw=40.0MiB/s, io=1200MiB, run=30001-30001msec

Both tests showed a reduction in latency of 25 microseconds.

Change-Id: I0045c3ec1ca42a72498287a5f828f3389a75e074